requirements.txt
glean-api-client>=0.1.0
mcp[cli]>=1.2.0
uvicorn>=0.30.0
